+++ controller/src/main/java/com/linbit/linstor/api/rest/v1/serializer/JsonGenTypes.java
@@ -1388,6 +1388,7 @@ public class JsonGenTypes
         public String passphrase;
         public String node_name;
         public boolean download_only = false;
+        public boolean force_restore = false;
     }
 
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
@@ -1418,6 +1419,7 @@ public class JsonGenTypes
         public String dst_stor_pool;
         public Map<String, String> stor_pool_rename = Collections.emptyMap();
         public boolean download_only = false;
+        public boolean force_restore = false;
         public boolean allow_incremental = true;
     }
 
@@ -1472,6 +1474,7 @@ public class JsonGenTypes
         public String node_name;
         public String dst_stor_pool;
         public Map<String, String> stor_pool_rename = Collections.emptyMap();
+        public boolean force_restore = false;
     }
 
     @JsonInclude(JsonInclude.Include.NON_EMPTY)